git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@3164 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/SyncFile.cs
@@ -26,8 +26,8 @@ using System.Threading;
 using System.Collections;
 using System.Security.Cryptography;
 using Simias.Storage;
-using Simias.Sync.Client;
 using Simias.Event;
+using Simias.Sync.Delta;
 
 namespace Simias.Sync
 {
@@ -99,9 +99,9 @@ namespace Simias.Sync
 		/// <summary>
 		/// Get the platform file handle.
 		/// </summary>
-		public IntPtr Handle
+		public FileStream outStream
 		{
-			get {return workStream.Handle;}
+			get {return workStream;}
 		}
 
 
@@ -204,8 +204,7 @@ namespace Simias.Sync
 		protected BaseFileNode	oldNode;
 		Exception				exception;
 		ManualResetEvent		asyncEvent = new ManualResetEvent(true);
-		
-		
+				
 		#endregion
 		
 		#region Constructor / Finalizer.
@@ -275,7 +274,7 @@ namespace Simias.Sync
 		/// <param name="count">The number of bytes to write.</param>
 		public void Copy(long originalOffset, long offset, int count)
 		{
-			int bufferSize = count > BlockSize ? BlockSize : count;
+			int bufferSize = count > HashData.BlockSize ? HashData.BlockSize : count;
 			byte[] buffer = new byte[bufferSize];
 
 			lock (this)
@@ -294,12 +293,21 @@ namespace Simias.Sync
 		}
 		
 		/// <summary>
-		/// Get the platform file handle.
+		/// Get the stream.
 		/// </summary>
-		public IntPtr Handle
+		public FileStream inStream
 		{
-			get {return workStream.Handle;}
+			get {return workStream;}
 		}
+
+		/// <summary>
+		/// Gets the original stream.
+		/// </summary>
+		public FileStream ReadStream
+		{
+			get {return stream;}
+		}
+
 
 		/// <summary>
 		/// Gets or Sets the file position.
@@ -324,9 +332,9 @@ namespace Simias.Sync
 		/// </summary>
 		public long Length
 		{
-			get { return stream == null ? 0 : stream.Length; }
+			get { return node.Length; }
 		}
-
+		
 		#endregion
 
 		#region protected methods.
@@ -337,7 +345,7 @@ namespace Simias.Sync
 		/// <param name="node">The node that represents the file.</param>
 		protected void Open(BaseFileNode node)
 		{
-			this.SetupFileNames(node, "");
+			SetupFileNames(node, "");
 			Log.log.Debug("Opening File {0}", file);
 			// Open the file so that it cannot be modified.
 			oldNode = collection.GetNodeByID(node.ID) as BaseFileNode;
@@ -348,7 +356,7 @@ namespace Simias.Sync
 			catch (FileNotFoundException)
 			{
 				// Check to see if we have a partially downloaded file to delta sync with.
-				if (File.Exists(workFile))
+				if (collection.Role == SyncCollectionRoles.Slave && File.Exists(workFile))
 				{
 					partialFile = workFile + ".part";
 					File.Move(workFile, partialFile);
@@ -486,8 +494,6 @@ namespace Simias.Sync
 		protected BaseFileNode	node;
 		/// <summary>The ID of the node.</summary>
 		protected string		nodeID;
-		/// <summary>The size of the Blocks that are hashed.</summary>
-		protected const int		BlockSize = 4096;
 		/// <summary>The maximun size of a transfer.</summary>
 		protected const int		MaxXFerSize = 1024 * 64;
 		/// <summary>The name of the actual file.</summary>
@@ -496,11 +502,12 @@ namespace Simias.Sync
 		protected string		workFile;
 		/// <summary>The Prefix of the working file.</summary>
 		const string			WorkFilePrefix = ".simias.wf.";
+		const string			MapFilePrefix = ".simias.map.";
 		static string			workBinDir = "WorkArea";
 		static string			workBin;
 		/// <summary>Used to publish Sync events.</summary>
-		static protected EventPublisher	eventPublisher = new EventPublisher();
-
+		static public			EventPublisher	eventPublisher = new EventPublisher();
+		
 		#endregion
 
 		#region protected methods.
@@ -531,6 +538,15 @@ namespace Simias.Sync
 					Directory.CreateDirectory(workBin);
 			}
 			this.workFile = Path.Combine(workBin, WorkFilePrefix + node.ID + sessionID);
+		}
+
+		/// <summary>
+		/// Retuns the name of the HashMap for this file.
+		/// </summary>
+		/// <returns></returns>
+		protected string GetMapFileName()
+		{
+			return Path.Combine(collection.ManagedPath, MapFilePrefix + node.ID);
 		}
 
 		#endregion
@@ -578,50 +594,4 @@ namespace Simias.Sync
 	}
 
 	#endregion
-
-	/// <summary>
-	/// Definitions for the http handler.
-	/// </summary>
-	public class SyncHttp
-	{
-		/// <summary>
-		/// 
-		/// </summary>
-		public static string	CopyOffset = "CopyOffset";
-		/// <summary>
-		/// 
-		/// </summary>
-		public static string	SyncRange = "SyncRange";
-		/// <summary>
-		/// 
-		/// </summary>
-		public static string	SyncOperation = "SyncOperation";
-		/// <summary>
-		/// 
-		/// </summary>
-		public static string	SyncBlocks = "SyncBlocks";
-		/// <summary>
-		/// 
-		/// </summary>
-		public static string	BlockSize = "SyncBlockSize";
-		/// <summary>
-		/// 
-		/// </summary>
-		public enum Operation
-		{
-			/// <summary>
-			/// 
-			/// </summary>
-			Read = 1,
-			/// <summary>
-			/// 
-			/// </summary>
-			Write,
-			/// <summary>
-			/// 
-			/// </summary>
-			Copy,
-		}
-	}
-	
 }
